In MetaFEM V0.1.0 there are in total 30 exposed API functions/macros/types/gloabal variables.

Core

MetaFEM has three basic data structures:

FEM_Domain, the overall system, defines $Kx=d$.

WorkPiece, a mesh assigned with physics, defines both domain physics and other numerical modifications like stabilization.

Boundary, an array of segment/face IDs assigned with physics, defines boundary condition.

MetaFEM.FEM_Domain — Type

FEM_Domain

A FEM_Domain contains everything needed to assemble a linear system Kx=d in FEM. The attributes are:

dim, the dimension, 2 or 3.

workpieces, the array of all the WorkPieces in this domain, which will be finally solved in fully coupling.

tools, reserved for the external geometry Tool, e.g., for contact. Not implemented.

time_discretization, the temporal discritization scheme. Currently only generalized-α method is implemented.

globalfield, the container for sparse K, dense x and d in Kx=d.

K_linear_func the generated function to update the linear part of K.

K_nonlinear_func the generated function to update the nonlinear part of K and the residue d.

linear_solver, the applied linear solver.

To add a FEM_Domain of dimension dim, the exposed API is:

FEM_Domain(; dim::Integer)

which returns the new FEM_Domain.

MetaFEM.WorkPiece — Type

WorkPiece

A WorkPiece is a meshed part assigned with some known physics. The attributes are:

ref_geometry, the first order mesh to describe the FEM_Geometry.

physics, the raw PDE weakforms.

local_assembly, the re-organized PDE weakforms with sorted/indexed variables.

max_sd_order, an explicit limit of the maximum spatial derivative order to save memory.

element_space, the information about the spatial discritization, i.e., interpolation and intergration.

mesh, the mesh regenerated according to the element_space and actually used in simulation.

To add a Workpiece with the geometry ref_geometry to the FEM_Domain fem_domain, the exposed API is:

add_WorkPiece(ref_geometry; fem_domain::FEM_Domain)

which returns the WorkPiece ID in the fem_domain.workpieces.

MetaFEM.add_Boundary — Function

add_Boundary(ID::Integer, bdy_ref_edge_IDs::CuVector; fem_domain::FEM_Domain = fem_domain, target::Symbol = :WorkPiece)

In a 2D/3D FEM_Domain fem_domain, mark the segment/face IDs of the WorkPiece fem_domain.workpieces[wp_ID] as a boundary, to assign physics later. Multiple boundaries are independent from each other and can share the same segment/face IDs. If target = :Tool, the boundary is of fem_domain.tools[wp_ID], not implemented.

The function returns the boundary group ID bg_ID.

Mesh

The input to define a mesh is (vert, connections), very similar to the mesh loading process in other fields, e.g., for plotting

vert is a $dim \times N_n$ CuArray of coordinates, where $N_n$ is the number of nodes.

connections is a $N_v \times N_e$ CuArray of sequenced node IDs where $N_v$ is the number of nodes in each element and $N_e$ is the total element number.

To define (vert, connections), we have the following helper functions:

MetaFEM.make_Square — Function

make_Square(x::Tuple, n::Tuple, shape = :CUBE)
make_Brick(x::Tuple, n::Tuple, shape = :CUBE)

Helper functions to creates a grid, i.e., the mesh nodes are positioned at cubic lattice points, with mesh shape = :CUBE or :SIMPLEX defining the connection.

MetaFEM.read_Mesh — Function

read_Mesh(filename::String)

The function loads an external mesh. Currently supported file types are:

.inp file, Abaqus first order mesh.

.mphtxt file, COMSOL mesh.

With (vert, connections), the first order mesh can be defined by:

MetaFEM.construct_TotalMesh — Function

construct_TotalMesh(coors::CuArray, connections::CuArray)

The function reads (vert, connections) and declares the first order mesh, FEM_Geometry.

There are 4 concepts in a FEM_Geometry:

Each vertex in vertices stores the vertex coordinates.

Each segment in segments stores the vertex IDs connected to this segment.

Each face in faces stores both the vertex IDs and the segment IDs connected to this face.

Each block in blocks stores the vertex IDs, the segment IDs and the face IDs connected to this block.

If applicable, the IDs are in order, e.g., segment IDs in a face are clockwise/counter-clockwise (since we can't define an outward normal for a bare face).

The function returns either a Geo_TotalMesh2D with attributes (vertices, segments, faces) or a Geo_TotalMesh3D with attributes (vertices, segments, faces, blocks).

We also have the helper function to define boundaries:

MetaFEM.get_BoundaryMesh — Function

get_BoundaryMesh(total_mesh::Geo_TotalMesh2D)
get_BoundaryMesh(total_mesh::Geo_TotalMesh3D)

Helper function to collect all the segment/face IDs which can be marked as boundary.

Symbolics

The APIs to define the physics are pretty straightforward:

MetaFEM.@Sym — Macro

@Sym (name, attribute_1, attribute_2,...)
@External_Sym (name, attribute_1, attribute_2,...)

@Sym declares a variable symbol name with the attributes attribute_1, attribute_2, ..., which is simply stored as the pair name => (attribute_1, attribute_2, ...) in the exposed global dictionary VARIABLE_ATTRIBUTES.

@External_Sym is just the @Sym with one default attribute: :EXTERNAL_VAR.

MetaFEM.@Def — Macro

@Def ex 
@Def begin
    ex1
    ex2
    ...
end

@Def defines a symbolic term, e.g., 1, σ{i, j}`, Bilinear(⋅, ⋅), etc.

MetaFEM.assign_Boundary_WeakForm — Function

assign_WorkPiece_WeakForm(wp_ID::Integer, this_term::SymbolicTerm; fem_domain::FEM_Domain)
assign_Boundary_WeakForm(wp_ID::Integer, bg_ID::Integer, this_term::SymbolicTerm; fem_domain::FEM_Domain)

The functions assign this_term, which is either a bilinear term Bilinear(⋅, ⋅), or a sum of the bilinear terms, to the target WorkPiece or boundary.

The internal mechanism is different from the classical approaches and worth mentioning: The expressions will go through symbolic rewriting with rewriting rules, enabling customizations.

Assembly

There are some necessary preprocesses:

MetaFEM.initialize_LocalAssembly — Function

initialize_LocalAssembly(fem_domain::FEM_Domain; explicit_max_sd_order::Integer = 9)
initialize_LocalAssembly(dim::Integer, workpieces::Vector{WorkPiece}; explicit_max_sd_order::Integer = 9)

This function preprocesses/reorganizes the weakforms. The input explicit_max_sd_order is the exposed API for explicitly limit high order spatial derivative.

MetaFEM.mesh_Classical — Function

mesh_Classical(wp_IDs; shape::Symbol, itp_type::Symbol = :Lagrange, itp_order::Integer, itg_order::Integer, fem_domain::FEM_Domain)

The function generates the mesh of order itp_order, interpolation type itp_type = :Lagrange/:Serendipity with gaussian quadrature of order itg_order on fem_domain.workpieces[wp_IDs]. The dimension and mesh type (:CUBE/:SIMPLEX) will follow the first order mesh of each WorkPiece.

MetaFEM.compile_Updater_GPU — Function

compile_Updater_GPU(; domain_ID::Integer, fem_domain::FEM_Domain)

The function generates fem_domain.K_linear_func and fem_domain.K_nonlinear_func. The input domain_ID is only used to generate the function name.

Update simulation

MetaFEM.update_Mesh — Function

update_Mesh(dim::Integer, wp::WorkPiece, this_space::Classical_Discretization)

This function updates Jacobians and interpolation values.

MetaFEM.assemble_Global_Variables — Function

assemble_Global_Variables(; fem_domain::FEM_Domain)

This function allocates the sparse K, dense x and d.

MetaFEM.update_OneStep — Function

update_OneStep(time_discretization::GeneralAlpha; max_iter::Integer = 4, fem_domain::FEM_Domain)

This function calculates K(Δx)=d and updates x += (Δx). max_iter determines the maximum iteration for Δx with different K and d in nonlinear cases. The converge tolerance is determined by fem_domain.globalfield.converge_tol while the linear solver is fem_domain.linear_solver.

MetaFEM.dessemble_X — Function

dessemble_X(workpieces::Vector{WorkPiece}, globalfield::GlobalField)

This function dessembles globalfield.x to local data, i.e., workpiece.mesh.controlpoint.sym, like workpiece.mesh.controlpoint.T.

Linear solvers

In update_OneStep, the linear solver, fem_domain.linear_solver is called, which is supposed to take a globalfield and return the dx. We provide the following linear solvers:

MetaFEM.solver_LU_CPU — Function

solver_LU_CPU(globalfield)

CPU LU solver.

MetaFEM.solver_QR — Function

solver_QR(globalfield; reorder::Integer = 1, singular_tol::Number)

GPU QR solver with CUDA's csrlsvqr!.

MetaFEM.solver_LU — Function

solver_LU_CPU(globalfield; reorder::Integer = 1, singular_tol::Number)

GPU LU solver with CUDA's csrlsvlu!. Note, csrlsvlu! is essentially a host function and the device part is not accessible.

MetaFEM.solver_BiCG — Function

solver_BiCG(globalfield; Pl_func::Function = precondition_Nothing, max_iter = 5000, max_pass = 4, l = 8)

GPU bicgstabl (a classical stablized bi conjugate gradient method) solver.

MetaFEM.solver_IDRs — Function

solver_IDRs(globalfield; Pl_func::Function = precondition_Nothing, max_iter = 5000, max_pass = 4, s = 8)

GPU IDRs (induced dimension reduction) solver.

with the following preconditioners:

MetaFEM.precondition_CUDA_Jacobi — Function

precondition_CUDA_Jacobi(A::CuSparseMatrixCSR{T})

GPU Jacobi preconditioner.

MetaFEM.precondition_CUDA_ILU — Function

precondition_CUDA_ILU(A)

GPU ILU preconditioner with CUDA's ilu02!.

Other helper functions

MetaFEM.write_VTK — Function

write_VTK(fname::String, wp::WorkPiece; scale = 1.)

Output all the variables in WorkPiece wp to file fname. The scale is only for the length scale, i.e., node coordinates.

MetaFEM.@Takeout — Macro

@Takeout a, b FROM c

equals

a = c.a
b = c.b

while

@Takeout a, b FROM c WITH PREFIX d

equals

da = c.a
db = c.b